Move ShowRunner and GadgetProvider out of Pinky.kt.
