+++ src/commonMain/kotlin/baaahs/Pinky.kt
@@ -3,14 +3,9 @@ package baaahs
 import baaahs.net.FragmentingUdpLink
 import baaahs.net.Network
 import baaahs.proto.*
-import baaahs.shaders.CompositingMode
-import baaahs.shaders.CompositorShader
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonConfiguration
-import kotlinx.serialization.serializer
 
 class Pinky(
     val sheepModel: SheepModel,
@@ -50,7 +45,7 @@ class Pinky(
         val gadgetProvider = GadgetProvider(pubSub)
 
         val buildShowRunner = {
-            ShowRunner(gadgetProvider, display, brains.values.toList(), beatProvider, dmxUniverse)
+            ShowRunner(gadgetProvider, brains.values.toList(), beatProvider, dmxUniverse)
         }
 
         var currentShowMetaData = selectedShow
@@ -145,131 +140,6 @@ class Pinky(
                 delay(delayTimeMillis.toLong())
             }
         }
-    }
-}
-
-class GadgetProvider(private val pubSub: PubSub.Server) {
-    val jsonParser = Json(JsonConfiguration.Stable)
-    private val activeGadgets = mutableListOf<GadgetData>()
-    private val activeGadgetChannel = pubSub.publish(Topics.activeGadgets, activeGadgets) {  }
-
-    private val gadgets = mutableMapOf<Gadget, GadgetChannel>()
-    private var nextGadgetId = 1
-
-    fun <G : Gadget> getGadget(gadget: G): G {
-        val gadgetId = nextGadgetId++
-
-        val topic =
-            PubSub.Topic("/gadgets/${gadget::class.simpleName}/$gadgetId", String.serializer())
-
-        val channel = pubSub.publish(topic, gadget.toJson().toString()) { updated ->
-            gadget.setFromJson(jsonParser.parseJson(updated))
-        }
-        gadgets[gadget] = GadgetChannel(topic, channel)
-
-        activeGadgets.add(GadgetData(gadget, topic.name))
-        activeGadgetChannel.onChange(activeGadgets)
-
-        return gadget
-    }
-
-    fun clear() {
-        gadgets.values.forEach { gadgetChannel -> gadgetChannel.channel.unsubscribe() }
-        gadgets.clear()
-        activeGadgets.clear()
-    }
-
-    class GadgetChannel(val topic: PubSub.Topic<String>, val channel: PubSub.Observer<String>)
-}
-
-class ShowRunner(
-    private val gadgetProvider: GadgetProvider,
-    private val pinkyDisplay: PinkyDisplay,
-    private val brains: List<RemoteBrain>,
-    private val beatProvider: Pinky.BeatProvider,
-    private val dmxUniverse: Dmx.Universe
-) {
-    private val brainsBySurface = brains.groupBy { it.surface }
-    private val shaderBuffers: MutableMap<Surface, MutableList<ShaderBuffer>> = hashMapOf()
-
-    fun getBeatProvider(): Pinky.BeatProvider = beatProvider
-
-    private fun recordShader(surface: Surface, shaderBuffer: ShaderBuffer) {
-        val buffersForSurface = shaderBuffers.getOrPut(surface) { mutableListOf() }
-
-        if (shaderBuffer is CompositorShader.Buffer) {
-            if (!buffersForSurface.remove(shaderBuffer.aShaderBuffer)
-                || !buffersForSurface.remove(shaderBuffer.bShaderBuffer)
-            ) {
-                throw IllegalStateException("Composite of unknown shader buffers!")
-            }
-        }
-
-        buffersForSurface += shaderBuffer
-    }
-
-    /**
-     * Obtain a shader buffer which can be used to control the illumination of a surface.
-     *
-     * @param surface The surface we're shading.
-     * @param shader The type of shader.
-     * @return A shader buffer of the appropriate type.
-     */
-    fun <B : ShaderBuffer> getShaderBuffer(surface: Surface, shader: Shader<B>): B {
-        val buffer = shader.createBuffer(surface)
-        recordShader(surface, buffer)
-        return buffer
-    }
-
-    /**
-     * Obtain a compositing shader buffer which can be used to blend two other shaders together.
-     *
-     * The shaders must already have been obtained using [getShaderBuffer].
-     */
-    fun getCompositorBuffer(
-        surface: Surface,
-        shaderBufferA: ShaderBuffer,
-        shaderBufferB: ShaderBuffer,
-        mode: CompositingMode = CompositingMode.OVERLAY,
-        fade: Float = 0.5f
-    ): CompositorShader.Buffer {
-        return CompositorShader(shaderBufferA.shader, shaderBufferB.shader)
-            .createBuffer(shaderBufferA, shaderBufferB)
-            .also {
-                it.mode = mode
-                it.fade = fade
-                recordShader(surface, it)
-            }
-    }
-
-    fun getDmxBuffer(baseChannel: Int, channelCount: Int) =
-        dmxUniverse.writer(baseChannel, channelCount)
-
-    fun getMovingHead(movingHead: SheepModel.MovingHead): Shenzarpy {
-        val baseChannel = Config.DMX_DEVICES[movingHead.name]!!
-        return Shenzarpy(getDmxBuffer(baseChannel, 16))
-    }
-
-    fun send(link: Network.Link) {
-        shaderBuffers.forEach { (surface, shaderBuffers) ->
-            if (shaderBuffers.size != 1) {
-                throw IllegalStateException("Too many shader buffers for $surface: $shaderBuffers")
-            }
-
-            val shaderBuffer = shaderBuffers.first()
-            val remoteBrains = brainsBySurface[surface]
-            remoteBrains?.forEach { remoteBrain ->
-                link.sendUdp(remoteBrain.address, Ports.BRAIN, BrainShaderMessage(shaderBuffer.shader, shaderBuffer))
-            }
-        }
-
-        dmxUniverse.sendFrame()
-    }
-
-    fun <T : Gadget> getGadget(gadget: T) = gadgetProvider.getGadget(gadget)
-
-    fun shutDown() {
-        gadgetProvider.clear()
     }
 }
